First Upload of Raspberry Pi code
This commit contains the line detection code for the Elcano Project. Within this commit is a README file which tells the user how to run the program, as well as code for Path Detection (which needs to be refined to avoid false positives), and recording videos with the Raspberry Pi.